Consolidated origin and direction generator interfaces with raster planner interface header
