+++ noether_tpp/tool_path_planners/include/noether_tpp/tool_path_planners/raster/raster_planner.h
@@ -28,7 +28,25 @@
 namespace noether_tpp
 {
 /**
- * @brief The RasterPlanner class - A specification of the tool path planner for covering surfaces
+ * @brief Interface for various methods to generate the direction of raster paths.  The direction generated will move along the line from the first point to the second, then third, etc.
+ */
+struct DirectionGenerator
+{
+  virtual ~DirectionGenerator() = 0;
+  virtual Eigen::Vector3d generate(const pcl::PolygonMesh& mesh) const = 0;
+};
+
+/**
+ * @brief Interface for setting the start point from which to generate toolpaths
+ */
+struct OriginGenerator
+{
+  virtual ~OriginGenerator() = default;
+  virtual Eigen::Vector3d generate(const pcl::PolygonMesh& mesh) const = 0;
+};
+
+/**
+ * @brief A specification of the tool path planner for covering surfaces
  * in repeating, evenly spaced path lines.  By default, a raster planner will produce a path with
  * all points having a consistent orientation, lines which all travel in the same direction across
  * the surface, and the traversal between lines will be in a consistent direction and order.
@@ -39,9 +57,7 @@ public:
   RasterPlanner(std::unique_ptr<DirectionGenerator>&& dir_gen, std::unique_ptr<OriginGenerator>&& origin_gen);
 
   /**
-   * @brief Creates a modifier that enforces the definition of a raster path on
-   * the path generated by a RasterPlanner.  It will enforce uniform waypoint orientation, uniform
-   * raster direction, and consistent raster ordering.
+   * @brief Creates a modifier that enforces the definition of a raster path on the path generated by a RasterPlanner.  It will enforce uniform waypoint orientation, uniform raster direction, and consistent raster ordering.
    * @return
    */
   static std::unique_ptr<const OneTimeToolPathModifier> createDefaultModifier();
